lib/python/requirements.txt
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0
openai>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
